+++ tests/runTests.bash
@@ -2,3 +2,4 @@
 
 ./phpunit.phar UnitTest lib/WsTmplTest.php
 ./phpunit.phar UnitTest lib/StreamsTest.php
+./phpunit.phar UnitTest lib/StreamsSearchIndexerTest.php